Fix data_returned and data_available
data_returned is defined as the total number of data that can be
returned using the passed parameters.
data_available is defined as the total number of data available of a
certain type at any time.

data_available is saved "lazily", i.e., when the first request has been
made, and then it is stored as a property for the specific collection
instance.

Tests have been slightly updated to account for the fix.
Perhaps the expected data_returned value should be hardcoded?