Add server_template.cfg in top-package dir
server_template.cfg is a template for a server.cfg file.
server.cfg is equivalent to an .ini file and can be read using
ConfigParser().
The point of server.cfg is to have a central place to write
server-specific paths to possibly changing configuration files.
At this point, it is used to direct the config.py towards the
config-file, whatever format it may be in, as well as pointing to the
index_links.json file that contains all resource objects to be added
under the index meta-database's /links endpoint.

If a server.cfg file does not exist upon loading CONFIG from config.py,
it will be created as a copy of server_template.cfg.