Remove old MongoTransformer
There is no need to have the old MongoTransformer.
The current MongoTransformer should work for the newest grammar always.
Either that or one should create a new Transformer per grammar version -
this may be desired.

Start to convert previously skipped tests to run, using `assertRaises`
for not yet implemented parts.