Add meta.description to BaseRelationshipResource
Sub-classed BaseResource to introduce special OPTiMaDe requirement of
having an OPTIONAL `meta` field in the relationship resource, which
MUST have the field `description`, it it's present.

Also, introduced some pylint disables for the models to avoid pylint
complaining about long lines and wrong use of `self` for validators
(which are class methods, hence they should NOT have the `self`
argument).