Introduce get_required_fields for mappers
The `handle_response_fields` utility function now assumes that all
fields to be excluded from entries are under `attributes`, since `id`,
`type` and all other top-level non-attributes fields are REQUIRED in the
response, even when not included as a value to `response_fields`.

The mappers can use the class variable `REQUIRED_FIELDS` to have a set
of REQUIRED response fields that MUST be returned, even when not
included as a value to `response_fields`.
Note: At the time of this commit, there are no fields under any
endpoint's attributes that are REQUIRED in the response.