Add `include` query parameter
Added to both multi and single entry query parameters.
A 400 Bad Request exception has been added, this may be collected in a
separate file with other custom exceptions that mainly sub-classes
FastAPI's `HTTPException`.

To provide an "empty" `include` parameter, the value has to be either ""
or ''.
If a resource not specified in the keys of
optimade/server/routers/__init__.py's ENTRY_COLLECTIONS is provided in
the value of `include` (and it's not an "empty" value) a 400 Bad Request
error will be raised and returned.