Reintroduce running run.sh index
There are now two ways to start the index meta-database server, instead
of the regular server.
1. Run: ./run.sh index
2. Set variable: export MAIN="main_index"
   And then run normally.
The two methods are also prioritized this way, i.e., if the environment
variable MAIN="main", but one runs ./run.sh index, then the index
meta-database will be started instead of the regular server.