Update tests for available_api_versions
Moved some of the validation for the fields "url" and "version" to the
Field.
For "url", a `regex` validator cannot be added in Field, since the type
"AnyHttpUrl" takes precedence, see
https://pydantic-docs.helpmanual.io/usage/schema/#unenforced-field-constraints
for more info.
Instead the validator is kept explicitly and the regex is added to the
OpenAPI schema as `pattern`.

Note the extra part of the version validator has been removed, where it
was checking whether or not each segment can be parsed as an integer.
This is now considered true always as a result of the regex validation.