Codecov-action supports token-less uploads
Removed secret token since the codecov-action now supports token-less
uploads for public GitHub repositories.
To make sure we only upload to Codecov for the upstream repository an
additional condition for when to upload is added to the upload step.

Added pydantic to requirements_eager.txt.

Using actions/checkout@master, since v2 is not up-to-date, and
codecov-action will not work for v2.
NOTE: This should be updated as soon as they have an up-to-date version.