Publish to PyPI after publishing release on GitHub
A python script checks the validity of the release metadata, i.e., if
the tag matches "vMAJOR.MINOR.PATCH", extras after PATCH are not
allowed, and the version _must_ be prefixed by "v".
Furthermore, it checks that the released tag points to a commit in the
`master` branch.

If nothing of these things are true, it will stop the publish on PyPI
workflow and one should delete the release from GitHub.