Add adapters - First adapter is for 'structures'
The goal of the adapters is to contain a conversion or adaptation
ability in a single Python class.

The class will be a proxy for `optimade.models.EntityResource`, where
the specific entity will take the place, e.g., `StructureResource`.

In this way, one can instatiate the adapter class with a JSON resource
object matching it, and get all the properties as attributes.
E.g., one could get the lattice vectors of a structure then by calling
`structures_adapter_class_object.attributes.lattice_vectors`.
The reason to not have the adapter class be a direct link or subclass of
the model class, is to not inherit all that comes with it, i.e., do not
inherit `pyndantic`'s `BaseModel`, etc.
While it can trip most IDEs, it is a more elegant solution that opens
the doors to other possibilities, in my opinion.

Besides being a handy wrapper for an entity resource object for a
client, it also serves as a converter/adapter to other known and widely
used libraries for the specific entity type.
For a structures resource object, this may be libraries such as
pymatgen, ASE, AiiDA, etc...
It may also be nice to be able to convert the structures resource object
to a Crystallographic Information File (CIF) file or other.
These adapter/conversion functions may be added in a separate file and
then referenced in the adapter class.

Thus, the adapter class stays small at first glance, but spreads its
tentacles wide throughout both the current package, but also a multitude
of other libraries, packages, and data formats.

In general, whenever the adapter class tries to get an attribute, it
goes through the following steps:

1. Try to fetch the OPTIMADE model property.
2. If the attribute starts with `get_`, try to look up if this is a
known conversion, then perform the conversion and return it as a string
or specific package/library Python class.
3. Raise AttributeError if the attribute is still unknown, i.e., the
previous steps failed and didn't raise already.