Add `sortable=True` to all properties
Since for MongoDB all properties can be sorted, it is done like this.
If one uses a different backend, one should overwrite the utility
function and use a different method to determine the sortability of
properties.

Slightly updated the wording of the `sortable` model attribute to be
more accurate: It is only a MUST if `sort` is supported AND the given
property can be used for sorting.