Big changes: Updating to v0.9.1 and v1.0.0-rc.2
Package version is made up to speed with the current latest release on
PyPI: v0.9.1
The OPTIMADE API specification is updated to the latest release
candidate version: v1.0.0-rc.2

The invoke tasks have been updated, so that nothing is raised, unless a
true error occurs.
If the package or OPTIMADE spec version is updated, the OpenAPI
specifications are also afterwards.

Instead of getting the package version from the installed version, it is
set through invoke tasks.
This means the `optimade/__init__.py` must be "manually" updated through
the invoke task "setver".