Add `api_hint` query parameter
This parameter is handled in a Middleware.
This is because it can lead to changing the original request by moving
to a versioned base URL.

We also need to include a new Exception, which is actually an HTTP
response, similar to BadRequest, namely VersionNotSupported (553).

The addition of the parameter to `query_params.py` is merely for adding
it to the OpenAPI specification.