Invoke task to create API Reference docs
This new task will go through the `optimade` folder and all subfolders,
excluding data folders.
A markdown file will be created for every python file and the folder
tree will be kept.
`.pages` files will also be created and `__init__.py` files will be
foregone, since the mkdocstrings plugin does not respect what can be
imported from the root of a module and what can not, it will simply add
everything with a docstring under the module ...

Due to the `__init__.py` file being foregone, the adapters that were
defined in said files have been moved to an `adapter.py` file.
No other `__init__.py` files define classes and functions that need to
be moved to separate files, so all good.

The new task has a flag `--pre-clean`, which will completely remove the
`api_reference` docs sub-folder and recreate all the files anew.
Otherwise the task will check the content of existing files to determine
whether it needs to rewrite them or skip writing.
This will of course not remove "old" files, for this the `--pre-clean`
flag should be added.