New middleware to catch any `OptimadeWarning`s
This is put as the first added middleware, since the order in which
middleware is added to the application matters.
See https://www.starlette.io/middleware/

The middleware is a bit hacky, since it's based on Starlette's
`BaseHTTPMiddleware`, which uses the `call_next` function that creates a
`StreamingResponse`.
This middleware will temporarily stitch the response back together, add
the warnings, and chop up the response again, returning a new
StreamingResponse.