Write out all mkdocstrings configuration
This is done to have easy access to any configurations should we want
it.
It also helps to show what the default is right away, instead of having
to check the mkdocstrings documentation.

Extend filters to exclude `__all__` and `__config__`.

Add to the invoke task that for `models` only, it should set
`rendering.show_if_no_docstring` to True.
This ensures we get _all_ models into the documentation, no matter if
the have doc strings or not.