Add extra css code
The code is lifted from the `mkdocstrings` repository's own
documentation.
It indents the references according to subsections and creates a
vertical gray line left of the text in the indentation.
Also, it makes sure not to break code weirdly.

Furthermore, the starting heading level is reset to the default (2).
And the `optimade` folder (and all sub-folders) are added to the watch
list for mkdocstrings to request a rebuild when using `mkdocs serve`
locally, whenever a file in `optimade/` is updated.