Use @ml-evs's trick with adapter dependencies
If an adapter dependency is not installed, a new type will be created
with the same name.
This is done for the sake of the docs.

In order to still warn and return about the non-existent dependency,
when trying to convert an OPTIMADE entry resource, the string
"optimade.adapters" is searched for in the representation of the "new"
type created.
This is because the new type will show up as belonging to the file in
which is was created.
This depends on a dependency never coming from within
"optimade.adapters", which makes sense, since it will always be external
libraries and packages.