Update `get_providers` with fallback
Order:

1. providers.optimade.org
2. GitHub source code raw content for providers repo
3. Log a warning

The list of providers has also been added to the index meta-database
implementation's `/links`-endpoint.
This required a purging of the MongoMock in-memory collections for the
invoke task that updates the OpenAPI JSON files, since the collection
names are shared, and the `insert_many` method would fail when importing
both applications without using re-initalizing the MongoMock in-memory
database and it's collections.