Update dependabot.yml for pip deps
Extend PRs to open to 50 (making it larger than the number of total
requirements in `requirements*.txt`.
Move to a weekly schedule of checking for dependencies.
I wanted bi-weekly, but this will have to do, since bi-weekly is not
possible.

Also, remove limit on open PRs for gitsubmodule to the default (5).