New CD workflow for updating `master` docs
This will deploy and update the `latest` version (alias `master`)
whenever a push event to `master` happens.
It will not update the CHANGELOG or other things. Indeed, it will fail
if the API Reference in the documentation has not been properly updated.

The "regular" CD workflow has been updated with versioned documentation
deployment, as well as a new feature:

The release description on GitHub will be updated with the changelog for
that particular version.
This changelog will be *appended* to the existing description.
One should now make sure to have a proper 1st level markdown header in
the description, since the appended changelog will introduce a similar
level header: `# Changelog`.