Down to 4 failing tests for the same thing
Lot's of changes here - mainly in
optimade.server.schemas:retrieve_queryable_properties().
Using the `model_fields` and FieldInfo instead of the generated JSON
Schema. Mainly due to the changes in the generated schema.

This led to a bit of hacking concerning the type (FieldInfo.annotation).
But otherwise, I think this is a better approach. Also implemented for
EntryCollection.get_attribute_fields().

Otherwise, changes here and there to make tests work as well as doing
some static typing sporadically.