Use 'groups' to group updates together in a single PR
This is for dependabot. Per package environment, i.e., for Python
dependencies and for GitHub Actions dependencies, a single PR will be
opened by dependabot, regularly, to update all dependencies within that
package environment.

This is similar, but not exactly the same, as what we have been doing
"manually" up to now via the 'dependabot_updates' branch.