Update release workflow
Use outputs whenever possible instead of setting environment variables
on-the-fly.

Use the docker image for github-changelog-generator straight from Docker
Hub, instead of using the CharMixer action.
This allows us to use the full toolbox available to us through the
github-changelog-generator software.

Create GitHub environments for 'pypi' and 'container-image' for the PyPI
release and the container image generation and deployment.
This results in building the distribution and publishing on PyPI having
its own job as well, which allows for the usage of Trusted Publishers.
For the container-image environment, this allows for an
environment-specific variable for the image name as well as keeping
track of the container image deployment.