Update dependency management and CI/CD workflows (#2087)
* Use 'groups' to group updates together in a single PR

This is for dependabot. Per package environment, i.e., for Python
dependencies and for GitHub Actions dependencies, a single PR will be
opened by dependabot, regularly, to update all dependencies within that
package environment.

This is similar, but not exactly the same, as what we have been doing
"manually" up to now via the 'dependabot_updates' branch.

* Remove the workflows for creating a single dependabot PR

This is now taken care of by dependabot itself. The 'dependabot_updates'
branch should be removed altogether.

* Update release workflow

Use outputs whenever possible instead of setting environment variables
on-the-fly.

Use the docker image for github-changelog-generator straight from Docker
Hub, instead of using the CharMixer action.
This allows us to use the full toolbox available to us through the
github-changelog-generator software.

Create GitHub environments for 'pypi' and 'container-image' for the PyPI
release and the container image generation and deployment.
This results in building the distribution and publishing on PyPI having
its own job as well, which allows for the usage of Trusted Publishers.
For the container-image environment, this allows for an
environment-specific variable for the image name as well as keeping
track of the container image deployment.

* Remove the auto-merge workflow

This would auto-merge PRs coming from Dependabot into the permanent
dependencies branch.

* Use outputs instead of env vars on-the-fly

Remove updating the 'dependabot_updates' branch

* Remove unused txt file and slightly update fly.toml

* Reorder package installation

Remove pre-commit job, instead use pre-commit.ci

* Add pre-commit.ci configuration

* Minimize usage of pre-commit.ci

Avoid autofixing PRs.
Skip running local hooks as they require the `invoke` package.
Return the custom pre-commit CI job, in order to ensure we run all
hooks.